Правки ResponseStatisticDto.

+++ stats/stats-dto/src/main/java/ru/practicum/DTO/ResponseStatisticDto.java
@@ -16,7 +16,7 @@ public class ResponseStatisticDto {
 
     private String uri;
 
-    private Integer hits;
+    private Long hits;
 }
